# Significant optimization of query test execution