Running `go test` in `query` directory is much to slow. After profiling, simple fix was applied to improve performance about 12 times (total time before change is about 60 seconds, and after about 5 seconds).
The most CPU consuming function in test is password encryption. But we can safely prepare encrypted passwords only for few tests. Running individual tests is also much faster (0.17s vs 0.01s).

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/1206)
<!-- Reviewable:end -->