@pawanrawal 2017-07-18
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> nice change 

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgraph-io/dgraph/1206#-:-KpKUVHENAMmrM4PcB1D:b-iu510k)*
<!-- Sent from Reviewable.io -->