@ashwin95r 2017-07-18
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 1 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgraph-io/dgraph/1206#-:-KpKUbbv9qH_RayxB8x-:bnfp4nl)*
<!-- Sent from Reviewable.io -->